requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Training utilities
tqdm>=4.65.0

# Web app (optional)
streamlit>=1.28.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
